# blog-main/src/main/java/com/longbai/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "微服务工程代码生成器,生成Dao、Service、Controller、Pojo、Feign、Swagger的JSON脚本",
    "version": "1.0.0",
    "title": "微服务工程代码生成器",
    "termsOfService": "https://blog.csdn.net/pkxwyf",
    "contact": {
      "email": "961968629@qq.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "petstore.swagger.io",
  "basePath": "/v2",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/tBlog": {
      "get": {
        "tags": [
          "TBlogController"
        ],
        "summary": "查询所有TBlog",
        "description": "查询所有TBlog方法详情",
        "operationId": "findAllUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "查询所有TBlog"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«TBlog»»"
            }
          }
        }
      },
      "post": {
        "tags": [
          "TBlogController"
        ],
        "summary": "添加TBlog",
        "description": "添加TBlog方法详情",
        "operationId": "addUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "添加TBlog方法详情",
            "required": true
              ,
              "schema": {
                "$ref": "#/definitions/TBlog"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "添加TBlog"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      }
    },
    "/tBlog/{id}": {
      "delete": {
        "tags": [
          "TBlogController"
        ],
        "summary": "根据ID删除TBlog",
        "description": "根据ID删除TBlog方法详情",
        "operationId": "deleteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID删除TBlog方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID删除TBlog"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "put": {
        "tags": [
          "TBlogController"
        ],
        "summary": "根据ID修改TBlog",
        "description": "根据ID修改TBlog方法详情",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改TBlog方法详情",
            "required": true
          },
          {
            "in": "body",
            "name": "TBlog",
            "description": "传入TBlog的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/TBlog"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改TBlog"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "get": {
        "tags": [
          "TBlogController"
        ],
        "summary": "根据ID查询TBlog",
        "description": "根据ID查询TBlog方法详情",
        "operationId": "findByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改TBlog方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改TBlog"
            ,
            "schema": {
              "$ref": "#/definitions/Result«TBlog»"
            }
          }
        }
      }
    },
    "/tBlog/search": {
      "post": {
        "tags": [
          "TBlogController"
        ],
        "summary": "不带分页条件搜索TBlog",
        "description": "不带分页条件搜索TBlog方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "TBlog",
            "description": "传入TBlog的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/TBlog"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索TBlog"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«TBlog»»"
            }
          }
        }
      }
    },
    "/tBlog/search/{page}/{size}": {
      "get": {
        "tags": [
          "TBlogController"
        ],
        "summary": "分页列表查询TBlog",
        "description": "分页列表查询TBlog方法详情",
        "operationId": "findPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索TBlog"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«TBlog»»"
            }
          }
        }
      },
      "post": {
        "tags": [
          "TBlogController"
        ],
        "summary": "分页条件搜索TBlog",
        "description": "分页条件搜索TBlog方法详情",
        "operationId": "findPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          },
          {
            "in": "body",
            "name": "TBlog",
            "description": "传入TBlog的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/TBlog"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "分页条件搜索TBlog"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«TBlog»»"
            }
          }
        }
      }
    },
...
    "/tComment/{id}": {
      "delete": {
        "tags": [
          "TCommentController"
        ],
        "summary": "根据ID删除TComment",
        "description": "根据ID删除TComment方法详情",
        "operationId": "deleteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID删除TComment方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID删除TComment"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "put": {
        "tags": [
          "TCommentController"
        ],
        "summary": "根据ID修改TComment",
        "description": "根据ID修改TComment方法详情",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改TComment方法详情",
            "required": true
          },
          {
            "in": "body",
            "name": "TComment",
            "description": "传入TComment的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/TComment"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改TComment"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "get": {
        "tags": [
          "TCommentController"
        ],
        "summary": "根据ID查询TComment",
        "description": "根据ID查询TComment方法详情",
        "operationId": "findByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改TComment方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改TComment"
            ,
            "schema": {
              "$ref": "#/definitions/Result«TComment»"
            }
          }
        }
      }
    },
    "/tComment/search": {
      "post": {
        "tags": [
          "TCommentController"
        ],
        "summary": "不带分页条件搜索TComment",
        "description": "不带分页条件搜索TComment方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "TComment",
            "description": "传入TComment的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/TComment"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索TComment"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«TComment»»"
            }
          }
        }
      }
    },
    "/tComment/search/{page}/{size}": {
      "get": {
        "tags": [
          "TCommentController"
        ],
        "summary": "分页列表查询TComment",
        "description": "分页列表查询TComment方法详情",
        "operationId": "findPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索TComment"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«TComment»»"
            }
          }
        }
      },
      "post": {
        "tags": [
          "TCommentController"
        ],
        "summary": "分页条件搜索TComment",
        "description": "分页条件搜索TComment方法详情",
        "operationId": "findPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          },
          {
            "in": "body",
            "name": "TComment",
            "description": "传入TComment的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/TComment"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "分页条件搜索TComment"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«TComment»»"
            }
          }
        }
      }
    },
...
    "/tImg/{id}": {
      "delete": {
        "tags": [
          "TImgController"
        ],
        "summary": "根据ID删除TImg",
        "description": "根据ID删除TImg方法详情",
        "operationId": "deleteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID删除TImg方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID删除TImg"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "put": {
        "tags": [
          "TImgController"
        ],
        "summary": "根据ID修改TImg",
        "description": "根据ID修改TImg方法详情",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改TImg方法详情",
            "required": true
          },
          {
            "in": "body",
            "name": "TImg",
            "description": "传入TImg的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/TImg"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改TImg"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "get": {
        "tags": [
          "TImgController"
        ],
        "summary": "根据ID查询TImg",
        "description": "根据ID查询TImg方法详情",
        "operationId": "findByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改TImg方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改TImg"
            ,
            "schema": {
              "$ref": "#/definitions/Result«TImg»"
            }
          }
        }
      }
    },
    "/tImg/search": {
      "post": {
        "tags": [
          "TImgController"
        ],
        "summary": "不带分页条件搜索TImg",
        "description": "不带分页条件搜索TImg方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "TImg",
            "description": "传入TImg的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/TImg"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索TImg"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«TImg»»"
            }
          }
        }
      }
    },
    "/tImg/search/{page}/{size}": {
      "get": {
        "tags": [
          "TImgController"
        ],
        "summary": "分页列表查询TImg",
        "description": "分页列表查询TImg方法详情",
        "operationId": "findPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索TImg"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«TImg»»"
            }
          }
        }
      },
      "post": {
        "tags": [
          "TImgController"
        ],
        "summary": "分页条件搜索TImg",
        "description": "分页条件搜索TImg方法详情",
        "operationId": "findPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          },
          {
            "in": "body",
            "name": "TImg",
            "description": "传入TImg的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/TImg"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "分页条件搜索TImg"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«TImg»»"
            }
          }
        }
      }
    },
...
    "/tTag/search": {
      "post": {
        "tags": [
          "TTagController"
        ],
        "summary": "不带分页条件搜索TTag",
        "description": "不带分页条件搜索TTag方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "TTag",
            "description": "传入TTag的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/TTag"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索TTag"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«TTag»»"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Result«List«TBlog»»": {
      "type": "object",
      "required": [
        "code",
        "data",
        "flag",
        "message"
      ],
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "返回状态码,20000:成功,20001:失败,20002:用户名或密码错误,20003:权限不足,20004:远程调用失败,20005:重复操作,20006:没有对应的数据"
        },
        "data": {
          "type": "array",
          "description": "逻辑数据",
          "items": {
            "$ref": "#/definitions/TBlog"
          }
        },
        "flag": {
          "type": "boolean",
          "example": false,
          "description": "执行是否成功,true:成功,false:失败"
        },
        "message": {
          "type": "string",
          "description": "提示信息"
        }
      },
      "description": "Result"
    },
...
    "TBlog": {
      "type": "object",
      "properties": {
        "blogId": {
          "type": "integer",
          "format": "int32",
          "description": "博文id"
        },
        "blogTitle": {
          "type": "string",
          "description": "博客标题"
        },
        "blogContent": {
          "type": "string",
          "description": "博客内容"
        },
        "userId": {
          "type": "integer",
          "format": "int32",
          "description": "用户id"
        },
        "typeId": {
          "type": "integer",
          "format": "int32",
          "description": "类型id"
        },
        "blogStatus": {
          "type": "integer",
          "format": "int32",
          "description": "博文状态0为草稿1为发布"
        },
        "includeImage": {
          "type": "string",
          "description": "文章中包含的图片url"
        },
        "coverImage": {
          "type": "string",
          "description": "封面图片url"
        },
        "createTime": {
          "type": "date",
          "description": "创建时间"
        },
        "updateTime": {
          "type": "date",
          "description": "更新时间"
        }
      },
      "description": "TBlog"
    },
...
  },
...
}